+++ crates/graph-utils/src/lib.rs
@@ -1,5 +1,7 @@
+mod graph_context;
 mod graph_formats;
 mod graph_traits;
 
+pub use graph_context::*;
 pub use graph_formats::*;
 pub use graph_traits::*;