+++ crates/task-expander/tests/token_expander_test.rs
@@ -18,8 +18,8 @@ mod token_expander {
         let sandbox = create_empty_sandbox();
         let project = create_project(sandbox.path());
         let task = create_task();
-        let context = create_context(&project, sandbox.path());
-        let expander = TokenExpander::new(context);
+        let context = create_context(sandbox.path());
+        let expander = TokenExpander::new(&project, &context);
 
         expander.replace_function(&task, "@unknown(id)").unwrap();
     }
@@ -30,8 +30,8 @@ mod token_expander {
         let sandbox = create_empty_sandbox();
         let project = create_project(sandbox.path());
         let task = create_task();
-        let context = create_context(&project, sandbox.path());
-        let expander = TokenExpander::new(context);
+        let context = create_context(sandbox.path());
+        let expander = TokenExpander::new(&project, &context);
 
         expander.replace_function(&task, "@files(unknown)").unwrap();
     }
@@ -44,8 +44,8 @@ mod token_expander {
         let sandbox = create_empty_sandbox();
         let project = create_project(sandbox.path());
         let task = create_task();
-        let context = create_context(&project, sandbox.path());
-        let expander = TokenExpander::new(context);
+        let context = create_context(sandbox.path());
+        let expander = TokenExpander::new(&project, &context);
 
         expander.replace_function(&task, "@in(str)").unwrap();
     }
@@ -56,8 +56,8 @@ mod token_expander {
         let sandbox = create_empty_sandbox();
         let project = create_project(sandbox.path());
         let task = create_task();
-        let context = create_context(&project, sandbox.path());
-        let expander = TokenExpander::new(context);
+        let context = create_context(sandbox.path());
+        let expander = TokenExpander::new(&project, &context);
 
         expander.replace_function(&task, "@in(10)").unwrap();
     }
@@ -70,8 +70,8 @@ mod token_expander {
         let sandbox = create_empty_sandbox();
         let project = create_project(sandbox.path());
         let task = create_task();
-        let context = create_context(&project, sandbox.path());
-        let expander = TokenExpander::new(context);
+        let context = create_context(sandbox.path());
+        let expander = TokenExpander::new(&project, &context);
 
         expander.replace_function(&task, "@out(str)").unwrap();
     }
@@ -82,8 +82,8 @@ mod token_expander {
         let sandbox = create_empty_sandbox();
         let project = create_project(sandbox.path());
         let task = create_task();
-        let context = create_context(&project, sandbox.path());
-        let expander = TokenExpander::new(context);
+        let context = create_context(sandbox.path());
+        let expander = TokenExpander::new(&project, &context);
 
         expander.replace_function(&task, "@out(10)").unwrap();
     }
@@ -98,8 +98,8 @@ mod token_expander {
             let mut task = create_task();
             task.inputs.push(InputPath::TokenFunc("@globs(all)".into()));
 
-            let context = create_context(&project, sandbox.path());
-            let expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let expander = TokenExpander::new(&project, &context);
 
             let result = expander.replace_function(&task, "@in(0)").unwrap();
 
@@ -118,8 +118,8 @@ mod token_expander {
             task.inputs
                 .push(InputPath::TokenFunc("@globs(unknown)".into()));
 
-            let context = create_context(&project, sandbox.path());
-            let expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let expander = TokenExpander::new(&project, &context);
 
             expander.replace_function(&task, "@in(0)").unwrap();
         }
@@ -132,8 +132,8 @@ mod token_expander {
             task.outputs
                 .push(OutputPath::TokenFunc("@globs(all)".into()));
 
-            let context = create_context(&project, sandbox.path());
-            let expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let expander = TokenExpander::new(&project, &context);
 
             let result = expander.replace_function(&task, "@out(0)").unwrap();
 
@@ -152,8 +152,8 @@ mod token_expander {
             task.outputs
                 .push(OutputPath::TokenFunc("@globs(unknown)".into()));
 
-            let context = create_context(&project, sandbox.path());
-            let expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let expander = TokenExpander::new(&project, &context);
 
             expander.replace_function(&task, "@out(0)").unwrap();
         }
@@ -173,8 +173,8 @@ mod token_expander {
             metadata.maintainers.push("user2".into());
 
             let task = create_task();
-            let context = create_context(&project, sandbox.path());
-            let expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let expander = TokenExpander::new(&project, &context);
 
             let get_value = |token: &str| {
                 expander
@@ -216,8 +216,8 @@ mod token_expander {
             );
 
             let task = create_task();
-            let context = create_context(&project, sandbox.path());
-            let expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let expander = TokenExpander::new(&project, &context);
 
             let get_value = |token: &str| {
                 expander
@@ -234,6 +234,8 @@ mod token_expander {
     }
 
     mod vars {
+        use env::consts;
+
         use super::*;
 
         #[test]
@@ -245,8 +247,8 @@ mod token_expander {
 
             let task = create_task();
 
-            let context = create_context(&project, sandbox.path());
-            let expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander
@@ -310,6 +312,12 @@ mod token_expander {
             );
             assert_eq!(
                 expander
+                    .replace_variable(&task, Cow::Borrowed("$workingDir"))
+                    .unwrap(),
+                sandbox.path().to_string_lossy()
+            );
+            assert_eq!(
+                expander
                     .replace_variable(&task, Cow::Borrowed("$workspaceRoot"))
                     .unwrap(),
                 sandbox.path().to_string_lossy()
@@ -346,6 +354,44 @@ mod token_expander {
                     .replace_variable(&task, Cow::Borrowed("$timestamp"))
                     .unwrap()
             ));
+
+            assert_eq!(
+                expander
+                    .replace_variable(&task, Cow::Borrowed("$arch"))
+                    .unwrap(),
+                consts::ARCH
+            );
+            assert_eq!(
+                expander
+                    .replace_variable(&task, Cow::Borrowed("$os"))
+                    .unwrap(),
+                consts::OS
+            );
+            assert_eq!(
+                expander
+                    .replace_variable(&task, Cow::Borrowed("$osFamily"))
+                    .unwrap(),
+                consts::FAMILY
+            );
+
+            assert_eq!(
+                expander
+                    .replace_variable(&task, Cow::Borrowed("$vcsBranch"))
+                    .unwrap(),
+                "master"
+            );
+            assert_eq!(
+                expander
+                    .replace_variable(&task, Cow::Borrowed("$vcsRepository"))
+                    .unwrap(),
+                "moonrepo/moon"
+            );
+            assert_eq!(
+                expander
+                    .replace_variable(&task, Cow::Borrowed("$vcsRevision"))
+                    .unwrap(),
+                "abcd1234"
+            );
         }
 
         #[test]
@@ -355,8 +401,8 @@ mod token_expander {
             project.language = LanguageType::JavaScript;
             let task = create_task();
 
-            let context = create_context(&project, sandbox.path());
-            let expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander
@@ -397,8 +443,8 @@ mod token_expander {
             project.language = LanguageType::JavaScript;
             let task = create_task();
 
-            let context = create_context(&project, sandbox.path());
-            let expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander
@@ -426,8 +472,8 @@ mod token_expander {
             let project = create_project(sandbox.path());
             let task = create_task();
 
-            let context = create_context(&project, sandbox.path());
-            let expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander
@@ -450,8 +496,8 @@ mod token_expander {
 
             task.command = "@files(sources)".into();
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             expander.expand_command(&task).unwrap();
         }
@@ -464,8 +510,8 @@ mod token_expander {
 
             task.command = "bin".into();
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(expander.expand_command(&task).unwrap(), "bin");
         }
@@ -478,8 +524,8 @@ mod token_expander {
 
             task.command = "$project/bin".into();
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(expander.expand_command(&task).unwrap(), "project/bin");
         }
@@ -492,8 +538,8 @@ mod token_expander {
 
             task.command = "$project/bin/$task".into();
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(expander.expand_command(&task).unwrap(), "project/bin/task");
         }
@@ -513,8 +559,8 @@ mod token_expander {
 
             task.command = "@meta(name)".into();
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(expander.expand_command(&task).unwrap(), "name");
         }
@@ -538,8 +584,8 @@ mod token_expander {
 
             task.args.push("@meta(name)".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(expander.expand_args(&task).unwrap(), vec!["name"]);
         }
@@ -553,8 +599,8 @@ mod token_expander {
             task.env.insert("FOO".into(), "bar".into());
             task.args.push("$FOO/$project".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(expander.expand_args(&task).unwrap(), vec!["bar/project"]);
         }
@@ -571,8 +617,8 @@ mod token_expander {
 
             task.env.insert("KEY".into(), "value".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_env(&task).unwrap(),
@@ -588,8 +634,8 @@ mod token_expander {
 
             task.env.insert("VAR".into(), "$project-prod".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_env(&task).unwrap(),
@@ -606,8 +652,8 @@ mod token_expander {
             task.env
                 .insert("VARS".into(), "$project-debug-$task".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_env(&task).unwrap(),
@@ -623,8 +669,8 @@ mod token_expander {
 
             task.env.insert("GROUP".into(), "@group(all)".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_env(&task).unwrap(),
@@ -643,8 +689,8 @@ mod token_expander {
 
             task.env.insert("DIRS".into(), "@dirs(dirs)".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_env(&task).unwrap(),
@@ -660,8 +706,8 @@ mod token_expander {
 
             task.env.insert("FILES".into(), "@files(all)".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_env(&task).unwrap(),
@@ -680,8 +726,8 @@ mod token_expander {
 
             task.env.insert("GLOBS".into(), "@globs(all)".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_env(&task).unwrap(),
@@ -697,8 +743,8 @@ mod token_expander {
 
             task.env.insert("ROOT".into(), "@root(all)".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_env(&task).unwrap(),
@@ -721,8 +767,8 @@ mod token_expander {
 
             task.env.insert("ROOT".into(), "@meta(name)".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_env(&task).unwrap(),
@@ -741,8 +787,8 @@ mod token_expander {
 
             task.env.insert("IN".into(), "@in(0)".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             expander.expand_env(&task).unwrap();
         }
@@ -758,8 +804,8 @@ mod token_expander {
 
             task.env.insert("OUT".into(), "@out(0)".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             expander.expand_env(&task).unwrap();
         }
@@ -775,8 +821,8 @@ mod token_expander {
 
             task.env.insert("OUT".into(), "@envs(envs)".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             expander.expand_env(&task).unwrap();
         }
@@ -793,8 +839,8 @@ mod token_expander {
 
             task.inputs = vec![InputPath::EnvVar("FOO_BAR".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_inputs(&task).unwrap(),
@@ -818,8 +864,8 @@ mod token_expander {
             env::set_var("FOO_THREE", "3");
             env::set_var("BAR_ONE", "1");
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             let mut result = expander.expand_inputs(&task).unwrap();
             result.env.sort();
@@ -846,8 +892,8 @@ mod token_expander {
 
             task.inputs = vec![InputPath::TokenFunc("@group(all)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_inputs(&task).unwrap(),
@@ -873,8 +919,8 @@ mod token_expander {
 
             task.inputs = vec![InputPath::TokenFunc("@dirs(dirs)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_inputs(&task).unwrap(),
@@ -897,8 +943,8 @@ mod token_expander {
 
             task.inputs = vec![InputPath::TokenFunc("@files(all)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_inputs(&task).unwrap(),
@@ -923,8 +969,8 @@ mod token_expander {
 
             task.inputs = vec![InputPath::TokenFunc("@globs(all)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_inputs(&task).unwrap(),
@@ -947,8 +993,8 @@ mod token_expander {
 
             task.inputs = vec![InputPath::TokenFunc("@root(all)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_inputs(&task).unwrap(),
@@ -968,8 +1014,8 @@ mod token_expander {
 
             task.inputs = vec![InputPath::TokenFunc("@envs(envs)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_inputs(&task).unwrap(),
@@ -991,8 +1037,8 @@ mod token_expander {
 
             task.inputs = vec![InputPath::TokenFunc("@in(0)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             expander.expand_inputs(&task).unwrap();
         }
@@ -1008,8 +1054,8 @@ mod token_expander {
 
             task.inputs = vec![InputPath::TokenFunc("@out(0)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             expander.expand_inputs(&task).unwrap();
         }
@@ -1025,8 +1071,8 @@ mod token_expander {
 
             task.inputs = vec![InputPath::TokenFunc("@meta(name)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             expander.expand_inputs(&task).unwrap();
         }
@@ -1042,8 +1088,8 @@ mod token_expander {
                 InputPath::TokenVar("$taskPlatform".into()),
             ];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_inputs(&task).unwrap(),
@@ -1071,8 +1117,8 @@ mod token_expander {
                 InputPath::WorkspaceGlob("cache/$target/files/**/*".into()),
             ];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_inputs(&task).unwrap(),
@@ -1102,8 +1148,8 @@ mod token_expander {
 
             task.inputs = vec![InputPath::ProjectFile("dir".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_inputs(&task).unwrap(),
@@ -1127,8 +1173,8 @@ mod token_expander {
 
             task.outputs = vec![OutputPath::TokenFunc("@group(all)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_outputs(&task).unwrap(),
@@ -1154,8 +1200,8 @@ mod token_expander {
 
             task.outputs = vec![OutputPath::TokenFunc("@dirs(dirs)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_outputs(&task).unwrap(),
@@ -1178,8 +1224,8 @@ mod token_expander {
 
             task.outputs = vec![OutputPath::TokenFunc("@files(all)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_outputs(&task).unwrap(),
@@ -1204,8 +1250,8 @@ mod token_expander {
 
             task.outputs = vec![OutputPath::TokenFunc("@globs(all)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_outputs(&task).unwrap(),
@@ -1228,8 +1274,8 @@ mod token_expander {
 
             task.outputs = vec![OutputPath::TokenFunc("@root(all)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_outputs(&task).unwrap(),
@@ -1252,8 +1298,8 @@ mod token_expander {
 
             task.outputs = vec![OutputPath::TokenFunc("@in(0)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             expander.expand_outputs(&task).unwrap();
         }
@@ -1269,8 +1315,8 @@ mod token_expander {
 
             task.outputs = vec![OutputPath::TokenFunc("@out(0)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             expander.expand_outputs(&task).unwrap();
         }
@@ -1286,8 +1332,8 @@ mod token_expander {
 
             task.outputs = vec![OutputPath::TokenFunc("@meta(name)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             expander.expand_outputs(&task).unwrap();
         }
@@ -1303,8 +1349,8 @@ mod token_expander {
 
             task.outputs = vec![OutputPath::TokenFunc("@envs(envs)".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             expander.expand_outputs(&task).unwrap();
         }
@@ -1322,8 +1368,8 @@ mod token_expander {
                 OutputPath::WorkspaceGlob("cache/$target/files/**/*".into()),
             ];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_outputs(&task).unwrap(),
@@ -1357,8 +1403,8 @@ mod token_expander {
                 OutputPath::WorkspaceGlob("cache/${BAR}/files/**/*".into()),
             ];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_outputs(&task).unwrap(),
@@ -1388,8 +1434,8 @@ mod token_expander {
 
             task.script = Some("bin --foo -az bar".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(expander.expand_script(&task).unwrap(), "bin --foo -az bar");
         }
@@ -1402,8 +1448,8 @@ mod token_expander {
 
             task.script = Some("$project/bin --foo -az bar".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_script(&task).unwrap(),
@@ -1419,8 +1465,8 @@ mod token_expander {
 
             task.script = Some("$project/bin/$task --foo -az bar".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_script(&task).unwrap(),
@@ -1437,8 +1483,8 @@ mod token_expander {
             task.script = Some("bin --foo -az @out(0)".into());
             task.outputs = vec![OutputPath::ProjectGlob("**/*.json".into())];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_script(&task).unwrap(),
@@ -1458,8 +1504,8 @@ mod token_expander {
                 InputPath::ProjectFile("other/file.json".into()),
             ];
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(
                 expander.expand_script(&task).unwrap(),
@@ -1482,8 +1528,8 @@ mod token_expander {
 
             task.script = Some("bin --name @meta(name)".into());
 
-            let context = create_context(&project, sandbox.path());
-            let mut expander = TokenExpander::new(context);
+            let context = create_context(sandbox.path());
+            let mut expander = TokenExpander::new(&project, &context);
 
             assert_eq!(expander.expand_script(&task).unwrap(), "bin --name name");
         }